Fix: 2 relevant lines not shown on website
The 2 lines:
let mut incoming = listener.incoming();
use futures::prelude::*; (or better just: use futures::StreamExt; )
are needed to compile the example but weren't published on the website. I removed the comments and change futures::prelude to use futures::StreamExt; 

As the linked example at the end is different, the 2 lines are needed to compile the code.