Make `Mutex` import explicit
The first time `std::sync::Mutex` is being imported and used during
`MiniTokio` exercise it's not explicit. Although Mutex is a well know
struct from the default sync lib, for a tutorial it makes more sense to
make it clear and avoid confusion. `std::sync::Arc` is also being silently
imported, but it was already used previously.

This commit makes the `std::sync::{Arc, Mutex}` import visible on the docs
removing the comment signal `#`.