Fix MiniTokio exercise code on tutorial
When following the tutorial [Async in depth][0] part, the resulted code is
missing some changes for a working code. It does not compile as the
`MiniTokio::new()` and `MiniTokio::spawn()` methods are not adjusted to
reflect the change from `VecDeque` to `crossbeam::channel`s.

This commit adds the missing changes and creates explanations for the added
code as part of the tutorial.

[0]: https://tokio.rs/tokio/tutorial/async