Async in depth tutorial improvements (#504)
* Make `Mutex` import explicit

The first time `std::sync::Mutex` is being imported and used during
`MiniTokio` exercise it's not explicit. Although Mutex is a well know
struct from the default sync lib, for a tutorial it makes more sense to
make it clear and avoid confusion. `std::sync::Arc` is also being silently
imported, but it was already used previously.

This commit makes the `std::sync::{Arc, Mutex}` import visible on the docs
removing the comment signal `#`.

* Fix MiniTokio exercise code on tutorial

When following the tutorial [Async in depth][0] part, the resulted code is
missing some changes for a working code. It does not compile as the
`MiniTokio::new()` and `MiniTokio::spawn()` methods are not adjusted to
reflect the change from `VecDeque` to `crossbeam::channel`s.

This commit adds the missing changes and creates explanations for the added
code as part of the tutorial.

[0]: https://tokio.rs/tokio/tutorial/async