bridging: Also discuss `spawn_blocking`
This page is currently mostly oriented towards discussing
how to handle a smaller asynchronous portion of a largely
synchronous code.

Users finding this page may be looking for the opposite case
which is discussed in the `spawn_blocking` docs.

There's obviously crossover here and we could discuss both
perhaps, but my initial goal here is to just cross-link things.