Added a small explanation of a change to Bytes (#666)
The "Update process()" section mentions that we indeed need to pass the `HashMap` as an argument. As I was going through those changes, the change from `Vec<u8>` to `Bytes` went under my radar (needs a change in line 138), and although it's simple enough to detect, I thought it would make sense to include this short explanation.