fix sed command to work with both GNU and MacOS sed
The sed command which replaces the markdown import in the api.js file
was not working on MacOS because the `-i` flag requires an argument
with the backup file extension. This commit adds the `.bak` extension
to the sed command to make it work on both GNU and MacOS sed (and then
removes the backup file).